Split up ContextExtensions into smaller files
